test for KanaOne
